Add PST-driven logging demo and preserve arena script
