+++ pst_tables.py
@@ -1,0 +1,75 @@
+# pst_tables.py
+# Базові (простi) Piece-Square Tables для мiдґейму.
+# Індексування під python-chess: 0=a1, ..., 63=h8 (білим знизу).
+# Для чорних використовуємо chess.square_mirror(sq).
+
+PST_MG = {
+    # PAWN
+    1: [
+        0,   0,   0,   0,   0,   0,   0,   0,
+        5,  10,  10, -20, -20,  10,  10,   5,
+        5,  -5, -10,   0,   0, -10,  -5,   5,
+        0,   0,   0,  20,  20,   0,   0,   0,
+        5,   5,  10,  25,  25,  10,   5,   5,
+        10, 10,  20,  30,  30,  20,  10,  10,
+        50, 50,  50,  50,  50,  50,  50,  50,
+        0,   0,   0,   0,   0,   0,   0,   0,
+    ],
+    # KNIGHT
+    2: [
+        -50,-40,-30,-30,-30,-30,-40,-50,
+        -40,-20,  0,  5,  5,  0,-20,-40,
+        -30,  5, 15, 20, 20, 15,  5,-30,
+        -30,  0, 20, 25, 25, 20,  0,-30,
+        -30,  0, 20, 25, 25, 20,  0,-30,
+        -30,  5, 15, 20, 20, 15,  5,-30,
+        -40,-20,  0,  5,  5,  0,-20,-40,
+        -50,-40,-30,-30,-30,-30,-40,-50,
+    ],
+    # BISHOP
+    3: [
+        -20,-10,-10,-10,-10,-10,-10,-20,
+        -10,  5,  0,  0,  0,  0,  5,-10,
+        -10, 10, 10, 10, 10, 10, 10,-10,
+        -10,  0, 10, 15, 15, 10,  0,-10,
+        -10,  5, 10, 15, 15, 10,  5,-10,
+        -10, 10, 10, 10, 10, 10, 10,-10,
+        -10,  5,  0,  0,  0,  0,  5,-10,
+        -20,-10,-10,-10,-10,-10,-10,-20,
+    ],
+    # ROOK
+    4: [
+          0,  0,  5, 10, 10,  5,  0,  0,
+         -5,  0,  0,  5,  5,  0,  0, -5,
+         -5,  0,  0,  5,  5,  0,  0, -5,
+         -5,  0,  0,  5,  5,  0,  0, -5,
+         -5,  0,  0,  5,  5,  0,  0, -5,
+         -5,  0,  0,  5,  5,  0,  0, -5,
+          5, 10, 10, 10, 10, 10, 10,  5,
+          0,  0,  5, 10, 10,  5,  0,  0,
+    ],
+    # QUEEN
+    5: [
+        -20,-10,-10, -5, -5,-10,-10,-20,
+        -10,  0,  5,  0,  0,  0,  0,-10,
+        -10,  5,  5,  5,  5,  5,  0,-10,
+         -5,  0,  5,  5,  5,  5,  0, -5,
+          0,  0,  5,  5,  5,  5,  0, -5,
+        -10,  0,  5,  5,  5,  5,  0,-10,
+        -10,  0,  0,  0,  0,  0,  0,-10,
+        -20,-10,-10, -5, -5,-10,-10,-20,
+    ],
+    # KING (мідґейм)
+    6: [
+        -30,-40,-40,-50,-50,-40,-40,-30,
+        -30,-40,-40,-50,-50,-40,-40,-30,
+        -30,-35,-40,-50,-50,-40,-35,-30,
+        -30,-35,-40,-50,-50,-40,-35,-30,
+        -20,-30,-35,-40,-40,-35,-30,-20,
+        -10,-20,-20,-20,-20,-20,-20,-10,
+         20, 20,  0,  0,  0,  0, 20, 20,
+         20, 30, 10,  0,  0, 10, 30, 20,
+    ],
+}
+
+PIECE_VALUES = {1: 100, 2: 320, 3: 330, 4: 500, 5: 900, 6: 0}
